refactor(kortixd): introduce internal harness service boundaries
